requirements.txt
discord.py
python-dotenv
streamlit
langchain
openai
wikipedia
chromadb
tiktoken
